Model selection and parameter estimation of MBM

Select the number of blocks per functional group using a stepwise search and estimate parameters

multipartiteBM(list_Net, v_distrib = NULL, namesFG = NULL,
  v_Kmin = 1, v_Kmax = 10, v_Kinit = NULL, initBM = FALSE,
  save = FALSE, verbose = TRUE, nbCores = NULL, maxiterVE = NULL)

Arguments

list_Net
A list of networks (defined via the function defineNetwork) i.e. multipartite networks

v_distrib
Type of proababilistic distributions in each network : if 0/1 then Bernoulli, if counting then Poisson. My default = Bernoulli. Must give a vector whose length is the number of networks in list_Net

namesFG
Names of functional groups (FG) (must correspond to names in list_Net)

v_Kmin
A vector of minimal number of blocks per functional group provided in the same order as in namesFG. v_Kmin can be a single value (same minimal number of blocks for all the FGs) or a vector with size equal to the number of FGs If v_Kmin is not specified, v_Kmin = 1.

v_Kmax
A vector of maximal number of blocks per functional group provided in the same order as in namesFG. v_Kmin can be a single value (same minimal number of blocks for all the FGs) or a vector with size equal to the number of FGs If v_Kmin is not specified, v_Kmin = 1.

v_Kinit
A vector of initial number of blocks per functional group provided in the same order as in namesFG. if v_Kinit is not specified, then several initialisations will be used : v_Kinit = v_Kmin, and v_Kinit = floor((v_Kmax + v_Kmin)/2)

initBM
If initBM = TRUE, then an aditional initialisation is done using simple LBM or SBM on each network separatly. The default value is FALSE

save
Set to TRUE to save the estimated parameters for intermediate visited models. Otherwise, only the better model (in ICL sense) is the ouput

verbose
Set to TRUE to display the current step of the search algorithm

nbCores
Number or cores used for the estimation. Not parallelized on windows. By default : half of the cores

maxiterVE
Maximum number of iterations if the VE step of the VEM algorithm. By default = 100

Value

a list of estimated parameters for the different models ordered by decreasing ICL. If save=FALSE, the length is of length 1

Examples

v_K <- c(3,2,2)
n_FG <- 3
list_pi <- vector("list", 3);
list_pi[[1]] <- c(0.4,0.3,0.3); list_pi[[2]] <- c(0.6,0.4); list_pi[[3]]  <- c(0.6,0.4)
E  = rbind(c(1,2),c(2,3),c(2,2))
v_distrib <- c('bernoulli','poisson','poisson')
typeInter <- c( "inc", "inc"  ,  "adj" )
list_theta <- list()
list_theta[[1]] <- matrix(rbeta(v_K[E[1,1]] * v_K[E[1,2]],1.5,1.5 ),nrow = v_K[E[1,1]], ncol = v_K[E[1,2]])
list_theta[[2]] <- matrix(rgamma(v_K[E[2,1]] * v_K[E[2,2]],7.5,1 ),nrow = v_K[E[2,1]], ncol = v_K[E[2,2]])
list_theta[[3]] <- matrix(rgamma(v_K[E[3,1]] * v_K[E[3,2]],7.5,1 ),nrow = v_K[E[3,1]], ncol = v_K[E[3,2]])
list_theta[[3]] <- 0.5*(list_theta[[3]] + t(list_theta[[3]])) # symetrisation for network 3
v_NQ = c(100,50,40)
dataSim <-  rMBM(v_NQ ,E , typeInter, v_distrib, list_pi, list_theta, seed=NULL, namesFG= c('A','B','D'),keepClassif = FALSE)
list_Net <- dataSim$list_Net
res <- multipartiteBM(list_Net, v_distrib = c("bernoulli","poisson","poisson"), namesFG = NULL, v_Kmin = 1,v_Kmax = 10,v_Kinit = NULL,verbose = TRUE, save=FALSE, maxiterVE = NULL)

#> [1] "------------Nb of entities in each functional group--------------"
#>   A   B   D 
#> 100  50  40 
#> [1] "------------Probability distributions on each network--------------"
#> [1] "bernoulli" "poisson"   "poisson"  
#> [1] "Searching the numbers of blocks starting from 1 1 1 clusters"
#> [1] "ICL : 14748.97 . Nb of clusters:  1 1 1"
#> [1] "ICL : 15132.3 . Nb of clusters:  1 1 2"
#> [1] "ICL : 15488.81 . Nb of clusters:  1 2 2"
#> [1] "ICL : 16190.21 . Nb of clusters:  2 2 2"
#> [1] "ICL : 16244.14 . Nb of clusters:  3 2 2"
#> [1] "Searching the numbers of blocks starting from 5 5 5 clusters"
#> [1] "ICL : 16009.56 . Nb of clusters:  5 5 5"
#> [1] "ICL : 16070.75 . Nb of clusters:  5 4 5"

#> Warning: Maximum number of VE iterations reached for model  4Maximum number of VE iterations reached for model  4Maximum number of VE iterations reached for model  5

#> Warning: Maximum number of VE iterations reached for model  5Maximum number of VE iterations reached for model  3Maximum number of VE iterations reached for model  5

#> [1] "ICL : 16127.61 . Nb of clusters:  3 3 4"
#> [1] "ICL : 16216.84 . Nb of clusters:  3 2 4"
#> [1] "ICL : 16229.96 . Nb of clusters:  3 2 3"
#> [1] "ICL : 16244.14 . Nb of clusters:  3 2 2"
#> [1] "Best model------ ICL : 16244.14 . Nb of clusters:  3 2 2 for A , B , D respectively"