Plot the mesoscopic view of the estimated MBM

#'

extractClustersMBM(fittedMBM, whichModel = 1)

Arguments

fittedMBM
A fitted Generalized BlockModel

whichModel
The index corresponding to the model to plot (default is 1, the best model)

Examples

npc1 <- 20 # number of nodes per block for functional group 1
Q1 <- 3 # blocks   for functional group 1
n1 <- npc1 * Q1 # number of nodes for functional group 1
Z1 <- diag(Q1)%x%matrix(1,npc1,1)
P1 <- matrix(runif(Q1*Q1),Q1,Q1)
A <- 1*(matrix(runif(n1*n1),n1,n1)<Z1%*%P1%*%t(Z1)) ## adjacency matrix
Agr <- defineNetwork(A,"diradj","FG1","FG1")  # First network
npc2 <- 40 #  number of nodes per block for functional group 2
Q2 <- 2 # blocks   for functional group 2
n2 <- npc2 * Q2 #  number of nodes for functional group 2
Z2 <- diag(Q2)%x%matrix(1,npc2,1)
P2 <- matrix(runif(Q1*Q2),Q1,Q2)
B <- 1*(matrix(runif(n1*n2),n1,n2)<Z1%*%P2%*%t(Z2)) ## incidence matrix
Bgr <- defineNetwork(B,"inc","FG1","FG2")
res <-   multipartiteBM(list(Agr,Bgr),namesFG = NULL,v_Kmin = 1,v_Kmax = 10,v_Kinit = NULL,verbose = TRUE, save=FALSE)

#> [1] "------------Nb of entities in each functional group--------------"
#> FG1 FG2 
#>  60  80 
#> [1] "------------Probability distributions on each network--------------"
#> NULL
#> [1] "Searching the numbers of blocks starting from 1 1 clusters"
#> [1] "ICL : -5617.91 . Nb of clusters:  1 1"
#> [1] "ICL : -5220.4 . Nb of clusters:  2 1"
#> [1] "ICL : -4875.44 . Nb of clusters:  3 1"
#> [1] "ICL : -4744.29 . Nb of clusters:  3 2"
#> [1] "Searching the numbers of blocks starting from 5 5 clusters"
#> [1] "ICL : -4921.02 . Nb of clusters:  5 5"
#> [1] "ICL : -4858.69 . Nb of clusters:  4 5"
#> [1] "ICL : -4817.6 . Nb of clusters:  3 5"
#> [1] "ICL : -4779.17 . Nb of clusters:  3 4"
#> [1] "ICL : -4758.77 . Nb of clusters:  3 3"
#> [1] "ICL : -4744.29 . Nb of clusters:  3 2"
#> [1] "Best model------ ICL : -4744.29 . Nb of clusters:  3 2 for FG1 , FG2 respectively"

extractClustersMBM (res,whichModel = 1)

#> $FG1
#> $FG1[[1]]
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20
#> 
#> $FG1[[2]]
#>  [1] 21 22 23 24 25 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40
#> 
#> $FG1[[3]]
#>  [1] 41 42 43 44 45 46 47 48 49 50 51 52 53 54 55 56 57 58 59 60
#> 
#> 
#> $FG2
#> $FG2[[1]]
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40
#> 
#> $FG2[[2]]
#>  [1] 41 42 43 44 45 46 47 48 49 50 51 52 53 54 55 56 57 58 59 60 61 62 63 64 65
#> [26] 66 67 68 69 70 71 72 73 74 75 76 77 78 79 80
#> 
#>