Model selection and parameter estimation of MBM

Select the number of blocks and identify the blocks per functional group using a variational EM algorithm

multipartiteBM(list_Net, v_distrib = NULL, namesFG = NULL,
  v_Kmin = 1, v_Kmax = 10, v_Kinit = NULL, initBM = TRUE,
  save = FALSE, verbose = TRUE, nbCores = NULL, maxiterVE = NULL,
  maxiterVEM = NULL)

Arguments

list_Net
a list of networks (defined via the function defineNetwork) i.e. a multipartite network

v_distrib
an optional vector of characters of length the number of networks and specifying the distribution used in each network (possible values bernoulli,poisson,gaussian,laplace)

namesFG
an optional vector of characters containing the names of functional groups (FG) (If Specified, must correspond to the names in list_Net)

v_Kmin
an optional vector of integers, specifying the minimal number of blocks per functional group (must be provided in the same order as in namesFG). v_Kmin may be a single value (same minimal number of blocks for all the FGs) or a vector with size equal to the number of FGs. Default value = 1.

v_Kmax
an optional vector of integers specifying the maximal number of blocks per functional group provided in the same order as in namesFG. v_Kmax may be a single value (same maximal number of blocks for all the FGs) or a vector with size equal to the number of FGs. Default value = 10.

v_Kinit
an optional vector of integers specifying initial numbers of blocks per FG provided in the same order as in namesFG. if v_Kinit is not specified, then v_Kinit = v_Kmin

initBM
an optional boolean. If initBM = TRUE an aditional initialisation is done using simple LBM or SBM on each network separatly. Default value = TRUE

save
an optional boolean. If TRUE save the estimated parameters for intermediate visited models. Otherwise, only the better model (in ICL sense) is the ouput. Default value = FALSE.

verbose
an optional boolean. If TRUE, display the current step of the search algorithm

nbCores
an optional integer specifying the number or cores used for the estimation. Not parallelized on windows. If ncores = NULL, then half of the cores are used.

maxiterVE
an optional integer specifying the maximum number of iterations in the VE step of the VEM algorithm. If NULL then default value = 1000

maxiterVEM
an optional integer specifying the maximum number of iterations of the VEM algorithm. If NULL then default value Default value = 1000

Value

a list of estimated parameters for the different models ordered by decreasing ICL. If save = FALSE, the length is of length 1 (only the better model is returned).

fittedModel

contains the results of the inference. res$fittedModel[[1]] is a list with fields
paramEstim

a MBMfit object.

ICL

the penalized likelihood criterion ICL.

vJ

the sequence of the varational bound of the likelihood through iterations of the VEM.

convergence

TRUE if the VEM reached convergence.

list_Net

contains the data.

Details

The function multipartiteBM selects the better numbers of blocks in each FG (with a penalized likelihood criterion). The model selection is performed with a forward backward strategy and the likelihood of each model is maximized with a variational EM).

Examples

v_K <- c(3,2,2)
n_FG <- 3
list_pi <- vector("list", 3);
list_pi[[1]] <- c(0.4,0.3,0.3); list_pi[[2]] <- c(0.6,0.4); list_pi[[3]]  <- c(0.6,0.4)
E  = rbind(c(1,2),c(2,3),c(2,2))
v_distrib <- c('bernoulli','poisson','poisson')
typeInter <- c( "inc", "inc"  ,  "adj" )
list_theta <- list()
list_theta[[1]] <- matrix(rbeta(v_K[E[1,1]] * v_K[E[1,2]],1.5,1.5 ),nrow = v_K[E[1,1]], ncol = v_K[E[1,2]])
list_theta[[2]] <- matrix(rgamma(v_K[E[2,1]] * v_K[E[2,2]],7.5,1 ),nrow = v_K[E[2,1]], ncol = v_K[E[2,2]])
list_theta[[3]] <- matrix(rgamma(v_K[E[3,1]] * v_K[E[3,2]],7.5,1 ),nrow = v_K[E[3,1]], ncol = v_K[E[3,2]])
list_theta[[3]] <- 0.5*(list_theta[[3]] + t(list_theta[[3]])) # symetrisation for network 3
v_NQ = c(100,50,40)
dataSim <-  rMBM(v_NQ ,E , typeInter, v_distrib, list_pi, list_theta, seed=NULL, namesFG= c('A','B','D'),keepClassif = FALSE)
list_Net <- dataSim$list_Net
res <- multipartiteBM(list_Net, v_distrib = c("bernoulli","poisson","poisson"))

#> [1] "------------Nb of entities in each functional group--------------"
#>   A   B   D 
#> 100  50  40 
#> [1] "------------Probability distributions on each network--------------"
#> [1] "bernoulli" "poisson"   "poisson"  
#> [1] "-------------------------------------------------------------------"
#> [1] " ------ Searching the numbers of blocks starting from [ 1 1 1 ] clusters"
#> [1] "ICL : 7531.51 . Nb of clusters: [ 1 1 1 ]"
#> [1] "ICL : 7795 . Nb of clusters: [ 1 2 1 ]"
#> [1] "ICL : 8280.31 . Nb of clusters: [ 2 2 1 ]"
#> [1] "ICL : 8322.18 . Nb of clusters: [ 3 2 1 ]"
#> [1] " ------ Searching the numbers of blocks starting from [ 3 2 1 ] clusters"
#> [1] "ICL : 8322.18 . Nb of clusters: [ 3 2 1 ]"
#> [1] "Best model------ ICL : 8322.18 . Nb of clusters: [ 3 2 1 ] for [ A , B , D ] respectively"